+++ src/main/kotlin/com/estapar/infrastructure/api/dto/GarageParkedEvent.kt
@@ -1,6 +1,6 @@
 package com.estapar.infrastructure.api.dto
 
-import com.estapar.domain.car.park.ParkAttempt
+import com.estapar.domain.car.park.attempt.ParkAttempt
 import com.fasterxml.jackson.annotation.JsonProperty
 
 open class GarageParkedEvent(
