+++ src/main/kotlin/com/estapar/infrastructure/component/ParkAttemptServiceImpl.kt
@@ -1,6 +1,6 @@
 package com.estapar.infrastructure.component
 
-import com.estapar.domain.car.park.ParkAttemptService
+import com.estapar.domain.car.park.attempt.ParkAttemptService
 import com.estapar.domain.car.park.ParkingService
 import com.estapar.domain.garage.spot.SpotService
 import org.springframework.stereotype.Service
